@@ -26,11 +26,17 @@
         }
     },
     "require": {
-        "php": ">=5.1.3",
+        "php": ">=7.4",
         "comodojo/exceptions": "dev-master"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.0|^5.0",
-        "scrutinizer/ocular": "^1.0"
+        "phpunit/phpunit": "^9.0",
+        "scrutinizer/ocular": "^1.8"
+    },
+    "minimum-stability": "dev",
+    "scripts": {
+        "test": "vendor/bin/phpunit --coverage-text --coverage-clover=coverage.clover",
+        "test-multi": "/usr/bin/bash .docker/test.sh",
+        "scrutinizer": "vendor/bin/ocular code-coverage:upload --format=php-clover coverage.clover"
     }
 }
