@@ -1,7 +1,9 @@
-<?php namespace Comodojo\Xmlrpc;
+<?php
+
+namespace Comodojo\Xmlrpc;
 
 use \Comodojo\Exception\XmlrpcException;
-use \Exception;
+use \SimpleXMLElement;
 
 /** 
  * XML-RPC decoder
@@ -21,61 +23,59 @@ use \Exception;
  * THE SOFTWARE.
  */
 
-class XmlrpcDecoder {
+class XmlrpcDecoder
+{
 
-    private $is_fault = false;
+    private bool $is_fault = false;
 
-    public function __construct() {
-
+    public function __construct()
+    {
         libxml_use_internal_errors(true);
-
     }
 
     /**
      * Decode an xmlrpc response
      *
      * @param   string  $response
-     *
      * @return  array
      *
-     * @throws  \Comodojo\Exception\XmlrpcException
+     * @throws  XmlrpcException
      */
-    public function decodeResponse($response) {
-
+    public function decodeResponse(string $response): array
+    {
         $xml_data = simplexml_load_string($response);
 
-        if ( $xml_data === false ) throw new XmlrpcException("Not a valid XMLRPC response");
+        if ($xml_data === false) {
+            throw new XmlrpcException("Not a valid XMLRPC response");
+        }
 
-        $data = array();
+        $data = [];
 
         try {
 
-            if ( isset($xml_data->fault) ) {
+            if (isset($xml_data->fault)) {
 
                 $this->is_fault = true;
-
                 array_push($data, $this->decodeValue($xml_data->fault->value));
+            } else if (isset($xml_data->params)) {
 
-            } else if ( isset($xml_data->params) ) {
-
-                foreach ( $xml_data->params->param as $param ) array_push($data, $this->decodeValue($param->value));
-
-            } else throw new XmlrpcException("Uncomprensible response");
-
+                foreach ($xml_data->params->param as $param) {
+                    array_push($data, $this->decodeValue($param->value));
+                }
+            } else {
+                throw new XmlrpcException("Uncomprensible response");
+            }
         } catch (XmlrpcException $xe) {
-            
+
             throw $xe;
-            
         }
 
-        return isset($data[0]) ? $data[0] : $data;
-
+        return $data[0] ?? $data;
     }
 
-    public function isFault() {
-
+    public function isFault()
+    {
         return $this->is_fault;
-
     }
 
     /**
@@ -87,284 +87,252 @@ class XmlrpcDecoder {
      * boxcarred request; a null value will be placed instead of array(method,params).
      *
      * @param   string  $request
-     *
      * @return  array
      *
-     * @throws  \Comodojo\Exception\XmlrpcException
+     * @throws  XmlrpcException
      */
-    public function decodeCall($request) {
-
+    public function decodeCall(string $request): array
+    {
         $xml_data = simplexml_load_string($request);
 
-        if ( $xml_data === false ) throw new XmlrpcException("Not a valid XMLRPC call");
-
-        if ( !isset($xml_data->methodName) ) throw new XmlrpcException("Uncomprensible request");
-            
-        $method_name = $this->decodeString($xml_data->methodName[0]);
-
-        if ( $method_name == "system.multicall" ) {
-
-            try {
-        
-                $data = $this->multicallDecode($xml_data);
-
-            } catch (XmlrpcException $xe) {
-                
-                throw $xe;
-
-            }
-
-        } else {
-
-            $parsed = array();
-
-            try {
-            
-                foreach ( $xml_data->params->param as $param ) $parsed[] = $this->decodeValue($param->value);
-
-            } catch (XmlrpcException $xe) {
-                
-                throw $xe;
-
-            }
-
-            $data = array($method_name, $parsed);
-
+        if ($xml_data === false) {
+            throw new XmlrpcException("Not a valid XMLRPC call");
         }
 
-        return $data;
+        if (!isset($xml_data->methodName)) {
+            throw new XmlrpcException("Uncomprensible request");
+        }
 
+        $method_name = $this->decodeString($xml_data->methodName[0]);
+
+        try {
+            if ($method_name === "system.multicall") {
+                return $this->multicallDecode($xml_data);
+            } else {
+                $parsed = [];
+                foreach ($xml_data->params->param as $param) {
+                    $parsed[] = $this->decodeValue($param->value);
+                }
+                return [$method_name, $parsed];
+            }
+        } catch (XmlrpcException $xe) {
+
+            throw $xe;
+        }
     }
 
     /**
      * Decode an xmlrpc multicall
      *
      * @param   string  $request
-     *
      * @return  array
      *
-     * @throws  \Comodojo\Exception\XmlrpcException
+     * @throws  XmlrpcException
      */
-    public function decodeMulticall($request) {
-
+    public function decodeMulticall(string $request): array
+    {
         $xml_data = simplexml_load_string($request);
 
-        if ( $xml_data === false ) throw new XmlrpcException("Not a valid XMLRPC multicall");
-
-        if ( !isset($xml_data->methodName) ) throw new XmlrpcException("Uncomprensible multicall request");
-
-        if ( $this->decodeString($xml_data->methodName[0]) != "system.multicall" ) throw new XmlrpcException("Invalid multicall request");
-
-        try {
-        
-            $data = $this->multicallDecode($xml_data);
-
-        } catch (XmlrpcException $xe) {
-            
-            throw $xe;
-
+        if ($xml_data === false) {
+            throw new XmlrpcException("Not a valid XMLRPC multicall");
         }
 
-        return $data;
+        if (!isset($xml_data->methodName)) {
+            throw new XmlrpcException("Uncomprensible multicall request");
+        }
 
+        if ($this->decodeString($xml_data->methodName[0]) != "system.multicall") {
+            throw new XmlrpcException("Invalid multicall request");
+        }
+
+        try {
+
+            return $this->multicallDecode($xml_data);
+        } catch (XmlrpcException $xe) {
+
+            throw $xe;
+        }
     }
 
     /**
      * Decode a value from xmlrpc data
      *
      * @param   mixed   $value
-     *
      * @return  mixed
      *
-     * @throws  \Comodojo\Exception\XmlrpcException
+     * @throws  XmlrpcException
      */
-    private function decodeValue($value) {
-
+    private function decodeValue($value)
+    {
         $children = $value->children();
 
-        if ( count($children) != 1 ) throw new XmlrpcException("Cannot decode value: invalid value element");
+        if (count($children) != 1) {
+            throw new XmlrpcException("Cannot decode value: invalid value element");
+        }
 
         $child = $children[0];
 
         $child_type = $child->getName();
 
-        switch ( $child_type ) {
+        switch ($child_type) {
 
             case "i4":
             case "int":
                 $return_value = $this->decodeInt($child);
-            break;
+                break;
 
             case "double":
                 $return_value = $this->decodeDouble($child);
-            break;
+                break;
 
             case "boolean":
                 $return_value = $this->decodeBool($child);
-            break;
+                break;
 
             case "base64":
                 $return_value = $this->decodeBase($child);
-            break;
-            
+                break;
+
             case "dateTime.iso8601":
                 $return_value = $this->decodeIso8601Datetime($child);
-            break;
+                break;
 
             case "string":
                 $return_value = $this->decodeString($child);
-            break;
+                break;
 
             case "array":
                 $return_value = $this->decodeArray($child);
-            break;
-            
+                break;
+
             case "struct":
                 $return_value = $this->decodeStruct($child);
-            break;
-            
+                break;
+
             case "nil":
             case "ex:nil":
                 $return_value = $this->decodeNil();
-            break;
-            
+                break;
+
             default:
                 throw new XmlrpcException("Cannot decode value: invalid value type");
-            break;
-
+                break;
         }
 
         return $return_value;
-
     }
 
     /**
      * Decode an XML-RPC <base64> element
      */
-    private function decodeBase($base64) {
-
-        return base64_decode($this->decodeString($base64));
-
+    private function decodeBase($field): string
+    {
+        return base64_decode($this->decodeString($field));
     }
 
     /**
      * Decode an XML-RPC <boolean> element
      */
-    private function decodeBool($boolean) {
-
-        return filter_var($boolean, FILTER_VALIDATE_BOOLEAN);
-
+    private function decodeBool($field): bool
+    {
+        return filter_var($field, FILTER_VALIDATE_BOOLEAN);
     }
 
     /**
      * Decode an XML-RPC <dateTime.iso8601> element
      */
-    private function decodeIso8601Datetime($date_time) {
-        
-        return strtotime($date_time);
-
+    private function decodeIso8601Datetime($field): int
+    {
+        return strtotime($field);
     }
 
     /**
      * Decode an XML-RPC <double> element
      */
-    private function decodeDouble($double) {
-
-        return (double) ($this->decodeString($double));
-
+    private function decodeDouble($field): float
+    {
+        return (float) ($this->decodeString($field));
     }
 
     /**
      * Decode an XML-RPC <int> or <i4> element
      */
-    private function decodeInt($int) {
-
-        return filter_var($int, FILTER_VALIDATE_INT);
-
+    private function decodeInt($field): int
+    {
+        return filter_var($field, FILTER_VALIDATE_INT);
     }
 
     /**
      * Decode an XML-RPC <string>
      */
-    private function decodeString($string) {
-
-        return (string) $string;
-
+    private function decodeString($field): string
+    {
+        return (string) $field;
     }
 
     /**
      * Decode an XML-RPC <nil/>
      */
-    private function decodeNil() {
-
+    private function decodeNil()
+    {
         return null;
-
     }
 
     /**
      * Decode an XML-RPC <struct>
      */
-    private function decodeStruct($struct) {
+    private function decodeStruct($field): array
+    {
+        $return_value = [];
 
-        $return_value = array();
+        foreach ($field->member as $member) {
 
-        foreach ( $struct->member as $member ) {
-
-            $name = $member->name."";
+            $name = (string) $member->name;
             $value = $this->decodeValue($member->value);
             $return_value[$name] = $value;
-
         }
 
         return $return_value;
-
     }
 
     /** 
      * Decode an XML-RPC <array> element
      */
-    private function decodeArray($array) {
+    private function decodeArray($field): array
+    {
+        $return_value = [];
 
-        $return_value = array();
-
-        foreach ( $array->data->value as $value ) {
+        foreach ($field->data->value as $value) {
 
             $return_value[] = $this->decodeValue($value);
-
         }
 
         return $return_value;
-
     }
 
     /** 
      * Decode an XML-RPC multicall request (internal)
-     * @param \SimpleXMLElement $xml_data
+     * 
+     * @param SimpleXMLElement $xml_data
      */
-    private function multicallDecode($xml_data) {
-
-        $data = array();
+    private function multicallDecode(SimpleXMLElement $xml_data): array
+    {
+        $data = [];
 
         try {
 
             $calls = $xml_data->params->param->value->children();
 
             $calls_array = $this->decodeArray($calls[0]);
-
-            foreach ( $calls_array as $call ) {
-                
-                $data[] = (!isset($call['methodName']) || !isset($call['params'])) ? null : array($call['methodName'], $call['params']);
-
+            foreach ($calls_array as $call) {
+                $data[] = (!isset($call['methodName']) || !isset($call['params']))
+                    ? null : [$call['methodName'], $call['params']];
             }
-            
         } catch (XmlrpcException $xe) {
-            
+
             throw $xe;
-
         }
-        
+
         return $data;
-
     }
-
 }
